+++ packages/implementations/react/src/atomic/Breadcrumbs/src/index.ts
@@ -1,5 +1,0 @@
-import { Breadcrumbs } from "./Breadcrumbs";
-
-export { Breadcrumbs } from "./Breadcrumbs";
-export type { BreadcrumbsProps } from "./breadcrumbs.types";
-export default Breadcrumbs;