+++ packages/implementations/react/src/composite/Breadcrumb/src/index.ts
@@ -1,0 +1,5 @@
+import { Breadcrumb } from "./Breadcrumb";
+
+export { Breadcrumb } from "./Breadcrumb";
+export type { BreadcrumbProps } from "./breadcrumb.types";
+export default Breadcrumb;