+++ packages/implementations/react/src/composite/Breadcrumb/webpack.config.ts
@@ -4,7 +4,7 @@ import { configuration } from "@rarui/webpack";
 const config = {
   output: {
     path: path.resolve(__dirname, "dist"),
-    library: "@rarui-react/breadcrumbs",
+    library: "@rarui-react/breadcrumb",
   },
 };