+++ packages/implementations/react/src/index.ts
@@ -23,6 +23,8 @@ export * from "@rarui-react/toggle";
 export * from "@rarui-react/tooltip";
 
 // composite
+export * from "@rarui-react/breadcrumb";
+export * from "@rarui-react/dropdown";
 export * from "@rarui-react/sidebar";
 export * from "@rarui-react/side-navigation";
 export * from "@rarui-react/stepper";