+++ packages/implementations/react/src/composite/Breadcrumb/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 Breadcrumbs consist of a list of links that help the user visualize the location of a page within the hierarchical structure of a website.
 
-## 2024-05-13 `1.0.0`
+## 2024-05-24 `1.0.0`
 
 #### 🎉 New features
 
-- Added `itens`, `maxItens` and `itensAfterTruncate` properties to the Component. ([#33](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/33) by [@mauricio.barbosa](https://git.rarolabs.com.br/mauricio.barbosa))
+- Added `children` and `itensAfterTruncate` properties to the Component. ([#33](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/33) by [@mauricio.barbosa](https://git.rarolabs.com.br/mauricio.barbosa))
 - Added stories on Component. ([#33](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/33) by [@mauricio.barbosa](https://git.rarolabs.com.br/mauricio.barbosa))
 
 <!-- #### 🛠 Breaking changes -->